# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Bagel's Repair Shop API"
  description: "API documentation for Bagel's Repair Shop"
  version: "0.0.1"
host: "127.0.0.1:5000" # Update with your server's host and port. It's in development right now
schemes:
  - "http"
consumes:
  - "application/json" #The API consumes JSON requests
produces:
  - "application/json" #The API produces JSON responses
securityDefinitions: #Defines the security scheme used in the API and how to add token authentication to endpoints
  bearerAuth: #name of the security scheme
    type: apiKey #type of security scheme
    name: Authorization
    in: header #where to find the token, in this case in the header of the request
paths: #Information about your API endpoints will go here
  /mechanics/login: #USER INPUT REQUIRED endpoint starts the path
    post: #request type (post, get, put, delete)
      tags:
        - Mechanics #adds this route to a category called Mechanics
      summary: "Mechanic login. Returns a JWT token upon successful authentication." #info about what this endpoint does
      description: "Authenticates a mechanic using email and password to receive a JWT token." #more detailed info
      parameters: #info about the input data. 
        - in: "body" #where the input data is located
          name: "body" #name of the input data
          description: "Mechanic login credentials" #info about the input data
          required: true #this input data is required
          schema:
            $ref: "#/definitions/MechanicLoginCredentials" #references the MechanicLoginCredentials definition below
      responses: #info about the output data
        200:
          description: "Successful login" #info about this response
          schema:
            $ref: "#/definitions/MechanicLoginResponse" #references the MechanicLoginResponse definition below
          examples:
            application/json:
              message: "Welcome John Doe"
              token: "aoj9q9214ntoiancp0qwr925n..."
        400:
          description: "Unauthorized - Invalid credentials" #info about this response
          schema:
            $ref: "#/definitions/400Response" #references the MechanicLoginResponse definition below
          examples:
            application/json:
              message: "Invalid email or password"
              token: null
  /mechanics: #USER INPUT REQUIRED
    post: #create mechanic route
      tags:
        - Mechanics
      summary: "Create a new mechanic"
      description: "Registers a new mechanic in the system."
      parameters:
        - in: "body"
          name: "body"
          description: "Mechanic registration details"
          required: true
          schema:
            $ref: "#/definitions/MechanicCreate"
      responses:
        201:
          description: "Mechanic created successfully"
          schema:
            $ref: "#/definitions/MechanicCreateResponse"
          examples:
            application/json:
              id: 1
              first_name: "John"
              last_name: "Doe"
              email: "mechanic@example.com"
              password: "hashedpassword"
              phone: "555-555-5555"
              speciality: "Engine Repair"
        400:
          description: "Bad Request - Invalid input data. Can be due to validation errors or duplicate email or phone."
          schema:
            $ref: "#/definitions/400Response"
          examples:
            application/json:
              message: "invalid input data or email/phone already exists"
               
    get: #get all mechanics route
      tags:
        - Mechanics
      summary: "Get all mechanics"
      description: "Retrieves a list of all mechanics in the system."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "List of mechanics retrieved successfully"
          schema:
              $ref: "#/definitions/GetAllMechanicsResponse"
          examples:
            application/json:
              - id: 1
                first_name: "John"
                last_name: "Doe"
                email: "example@gmail.com"
                phone: "555-555-5555"
                specialty: "Engine Repair"
        403:
          description: "Forbidden - Missing or invalid JWT token."
          examples:
            application/json:
              message: "Missing or invalid token"
        404:
          description: "Not Found - No mechanics found in the system."
          examples:
            application/json:
              message: "No mechanics found"

    put: #update mechanic route
      tags:
        - Mechanics
      summary: "Update mechanic details"
      description: "Updates the details of an existing mechanic. User will be identified via the JWT token."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Updated mechanic details"
          required: true
          schema:
            $ref: "#/definitions/MechanicCreate"
      responses:
        200:
          description: "Mechanic updated successfully"
          schema:
            $ref: "#/definitions/MechanicCreateResponse"
        400:
          description: "Bad Request - Invalid input data."
        403:
          description: "Forbidden - Missing or invalid mechanic JWT token."
        404:
          description: "Not Found - Mechanic not found."
  /mechanics/{mechanic_id}: #TOKEN REQUIRED delete mechanic route
    delete:
      tags:
        - Mechanics
      summary: "Delete mechanic account"
      description: "Deletes the mechanic's account from the system. User will be identified via the JWT token and can only delete their account."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Mechanic deleted successfully"
          examples:
            application/json:
              message: "Mechanic account deleted successfully"
        403:
          description: "Forbidden - Missing or invalid mechanic JWT token."
          schema:
            $ref: "#/definitions/403MechanicResponse"
          examples:
            application/json:
              message: "Missing or invalid token"
        404:
          description: "Not Found - Mechanic not found."
          schema:
            $ref: "#/definitions/404MechanicResponse"
          examples:
            application/json:
              message: "Mechanic not found"
    get:
      tags:
        - Mechanics
      summary: "Get mechanic by ID"
      description: "Retrieves the details of a specific mechanic by their ID."
      security:
        - bearerAuth: []
      parameters:
        - name: "mechanic_id"
          in: "path"
          description: "ID of the mechanic to retrieve"
          required: true
          type: "integer"
      responses:
        200:
          description: "Mechanic details retrieved successfully"
          schema:
            $ref: "#/definitions/MechanicCreateResponse"
        403:
          description: "Forbidden - Missing or invalid mechanic JWT token."
          schema:
            $ref: "#/definitions/403MechanicResponse"
        404:
          description: "Not Found - Mechanic not found."
          schema:
            $ref: "#/definitions/404MechanicResponse"
  /mechanics/my_tickets: #TOKEN REQUIRED get tickets assigned to logged-in mechanic
    get:
      tags:
        - Mechanics
      summary: "Get tickets assigned to the logged-in mechanic"
      description: "Retrieves all tickets assigned to the authenticated mechanic."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "List of tickets retrieved successfully"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/definitions/ServiceTicketResponse"
        403:
          description: "Forbidden - Missing or invalid mechanic JWT token."
        404:
          description: "Not Found - No tickets found for the mechanic."
  /mechanics/logout: #TOKEN REQUIRED mechanic logout route
    post:
      tags:
        - Mechanics
      summary: "Mechanic logout"
      description: "Logs out the authenticated mechanic by invalidating their JWT token."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Mechanic logged out successfully"
          examples:
            application/json:
              message: "Successfully logged out"
        403:
          description: "Forbidden - Missing or invalid mechanic JWT token."
          examples:
            application/json:
              message: "Missing or invalid token"
...
definitions: ##Information about the shape of input and output json
  MechanicLoginCredentials: #input data for login routes
    type: object
    properties:
      email:
        type: string
        example: "user@example.com"
      password:
        type: string
        example: "securepassword"

    required:
      - email
      - password
  MechanicLoginResponse: #output data for login routes
    type: object
    properties:
      message:
        type: string
      token:
        type: string
  MechanicCreate: #input data for create mechanic route
    type: object
    properties:
      first_name:
        type: string
        example: "John"
      last_name:
        type: string
        example: "Doe"
      password:
        type: string
        example: "securepassword"
      email:
        type: string
        example: "mechanic@example.com"
      phone:
        type: string
        example: "555-555-5555"
      specialty:
        type: string
        example: "Engine Repair" 
    required:
      - first_name
      - last_name
      - password
      - email
      - phone
      - specialty
  MechanicCreateResponse: #output data for create mechanic route
    type: object
    properties:
      id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      email:
        type: string
      phone:
        type: string
      specialty:
        type: string
        example: "Engine Repair" 
      role:
        type: string
        example: "mechanic"
  GetAllMechanicsResponse:
    type: array
    items:
      type: object
      properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: string
        specialty:
          type: string
...
  ServiceTicketResponse:
    type: array
    items:
      type: object
      properties:
        id:
          type: integer
        vehicle_make:
          type: string
        vehicle_model:
          type: string
        vehicle_year:
          type: integer
        service_description:
          type: string
        date_created:
          type: string
          format: date-time
        price:
          type: number
          format: float
        status:
          type: string
        parts:
          type: string
        mechanic_id: #null until assigned
          type: integer
        customer_id:
          type: integer
...
  400Response: #generic 400 response
    type: object
    properties:
      message:
        type: string
        example: "invalid input data or email/phone already exists"
  403MechanicResponse: #generic 403 response 
    type: object
    properties:
      message:
        type: string
        example: "Missing or invalid token"
...
  404MechanicResponse: #generic 404 response
    type: object
    properties:
      message:
        type: string
        example: "Mechanic not found"
...